Module Conduit_async.Service

type ('cfg, 't, 'flow) impl = (module SERVICE with type configuration = 'cfg and type flow = 'flow and type t = 't)

type ('cfg, 't, 'flow) service

The type for services, e.g. service-side protocols. 'cfg is the type for configuration, 't is the type for state states. 'flow is the type for underlying flows.

val equal : ('cfg0, 't0, 'flow0) service -> ('cfg1, 't1, 'flow1) service -> (('cfg0, 'cfg1) Conduit.refl * ('t0, 't1) Conduit.refl * ('flow0, 'flow1) Conduit.refl) option

equal svc0 svc1 proves that svc0 and svc1 are physically the same. For instance, Conduit asserts:

let service = Service.register ~service:(module V) ;;

let () = match Service.equal service service with
  | Some (Refl, Refl, Refl) -> ...
  | _ -> assert false

val register : service:('cfg, 't, 'flow) impl -> ('cfg, 't, 'flow) service

register ~service is the service using the implementation service. service must define make and accept function to be able to create server-side flows.

For instance:

module TCP : SERVICE with type configuration = Unix.sockaddr
                      and type t = Unix.file_descr
                      and type flow = Unix.file_descr

let tcp_service : (Unix.sockaddr, Unix.file_descr, Unix.file_descr) Service.service =
  Service.register ~service:(module TCP)

type error = [

| `Msg of string

]

val pp_error : error Fmt.t

val init : 'cfg -> service:('cfg, 't, 'flow) service -> ('t, [> error ]) Stdlib.result io

init cfg ~service initialises the service with the configuration cfg.

val accept : service:('cfg, 't, 'flow) service -> 't -> ('flow, [> error ]) Stdlib.result io

accept service t waits for a connection on the service t. The result is a flow connected to the client.

val close : service:('cfg, 't, 'flow) service -> 't -> (unit, [> error ]) Stdlib.result io

close ~service t releases the resources associated to the server t.

val impl : ('cfg, 't, 'flow) service -> (module SERVICE with type configuration = 'cfg and type flow = 'flow and type t = 't)

impl service is service's underlying implementation.