Module type Conduit_async.PROTOCOL

A protocol is a FLOW plus connect.

include Conduit__.Sigs.FLOW

type +'a io

type flow

Input & Output.

Depending on the I/O model, the type for inputs and outputs can differ ; for instance they could allow users the ability to define capabilities on them such as read or write capabilities.

However, in most of the current Conduit backends:

type input = Cstruct.t
type output = Cstruct.t

type input

and output

Errors.

type error

The type for errors.

val pp_error : error Fmt.t

pp_error is the pretty-printer for error.

val recv : flow -> input -> (int Conduit__.Sigs.or_end_of_flow, error) Stdlib.result io

recv flow input is Ok (`Input len) iff len bytes of data has been received from the flow flow and copied in input.

val send : flow -> output -> (int, error) Stdlib.result io

send t output is Ok len iff len bytes of data from output has been sent over the flow flow.

val close : flow -> (unit, error) Stdlib.result io

close flow closes flow. Subsequent calls to recv on flow will return `End_of_flow. Subsequent calls to send on t will return an Error.

type endpoint

val connect : endpoint -> (flow, error) Stdlib.result io